+++ src/tools/fuzzing.h
@@ -29,6 +29,7 @@ high chance for set at start of loop
 #include <ir/find_all.h>
 #include <ir/literal-utils.h>
 #include <ir/manipulation.h>
+#include "ir/memory-utils.h"
 #include <ir/utils.h>
 
 namespace wasm {
@@ -254,9 +255,7 @@ private:
   }
 
   void setupMemory() {
-    wasm.memory.exists = true;
-    // use one page
-    wasm.memory.initial = wasm.memory.max = 1;
+    MemoryUtils::ensureExists(wasm.memory);
     if (features.hasBulkMemory()) {
       size_t memCovered = 0;
       // need at least one segment for memory.inits