+++ test/lld/shared.c
@@ -1,0 +1,7 @@
+int puts(const char* str);
+extern int external_var;
+
+int print_message() {
+  puts("Hello, world");
+  return external_var;
+}